requirements.txt
discord.py
discord
stoat.py
requests